Optimised performance of getActivity.php by pre-processing the m_log in moodle into a new table called m_log_summary that aggregates student activity from the m_log.
The query to create m_log_summary is found in doc/m_log_summary.sql along with a new index and example queries to retrieve clicks and distinct students using m_log_summary.

The old queries that worked directly on the m_log are also given in m_log_summary.sql for posterity and in case need to reimplement for querying specific weeks.

The getActivity.php was updated to now work with the changed getActivityQueryTemplate.php script which generates the new queries for m_log_summary