Added pagelink attribute to the json data sent to the getActivity.php script so that the getActivity.txt log file knows what page was viewed in Moodle.
Fixed bug in the GM_XHR.js file which did not call that.onreadystatechange() when the onerror method is called.  If an error occurred with an ajax call using greasemonkey, the no handlers are called, and so the spinning busy wheel remains on the screen.  This fixes it so that the ajax call completes and the spinning wheel is removed.  This is important for times when for instance the web server might be down.