Added the shared Term class to staffUsage.php so that it can properly show usage statistics according to weeks of term.
Added 'types' option to staffUsage.php so that an additional report can be run showing the number views for different page types in Moodle. eg. book, page, quiz, assignment, course.

Improved commenting in some parts of the script

For the weeks usage report, changed to csv output format and added column for number of staff who used MAV for the given week.

Changed the week output format from the week code to the week name.